% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMED.R
\name{SMED}
\alias{SMED}
\title{SMED}
\usage{
SMED(f, p, n = 10, nc = 100, max.time = NULL, X0 = NULL, Xopt = NULL)
}
\arguments{
\item{f}{Function}

\item{p}{Number of dimensions}

\item{n}{Number of points}

\item{nc}{Number of points for contour plot}

\item{max.time}{max.time for GenSA optimization for each point}

\item{X0}{Matrix of initial points}

\item{Xopt}{Matrix of candidate points}
}
\value{
Points selected
}
\description{
SMED
}
\examples{
SMED(function(xx){xx[1]+xx[2]^2-sin(2*pi*xx[3])},p=3,n=10,max.time=.2)
}
