% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMED_1D.R
\name{SMED_1D}
\alias{SMED_1D}
\title{SMED in 1 dimension}
\usage{
SMED_1D(f, n = 10, navail = 1000, kk = 4, xmin = 0, xmax = 1,
  include.histogram = F)
}
\arguments{
\item{f}{Function}

\item{n}{Number of points}

\item{navail}{Number of candidate points}

\item{kk}{Power to use for MED}

\item{xmin}{Min x value}

\item{xmax}{Max x value}

\item{include.histogram}{Should histogram be plotted at end}
}
\value{
Nothing, plots points
}
\description{
SMED in 1 dimension
}
\examples{
SMED_1D(function(xx)dbeta(xx,4,2),n=30)
}
