% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMED_2D.R
\name{SMED_2D}
\alias{SMED_2D}
\title{SMED 2D}
\usage{
SMED_2D(f, n = 10, max.time = NULL)
}
\arguments{
\item{f}{Function}

\item{n}{Number of points to select}

\item{max.time}{max.time for GenSA optimization for each point}
}
\value{
Matrix of points
}
\description{
SMED 2D
}
\examples{
SMED_2D(TestFunctions::banana, 10, .1)
}
