% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMED_GP.R
\name{SMED_GP}
\alias{SMED_GP}
\title{SMED using Gaussian process models}
\usage{
SMED_GP(f, p, n0 = 10, n = 10, GP.package = "", opt.method = "genoud",
  contour.fit = 0, continue.option = F)
}
\arguments{
\item{f}{Function}

\item{p}{# of dimensions}

\item{n0}{Number of initial points from LHS}

\item{n}{Number of points to select.}

\item{GP.package}{R package to use for GP}

\item{opt.method}{Optimization method, either 'genoud' or 'LHS}

\item{contour.fit}{How often to refit contour plot}

\item{continue.option}{Whether option to continue should be given}
}
\value{
X design matrix and Y function values
}
\description{
SMED using Gaussian process models
}
\examples{
SMED_GP(f=TestFunctions::banana, p=2, n0=20, n=10, GP.package='laGP')
}
