% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMED_GPE_2D.R
\name{SMED_GPE_2D}
\alias{SMED_GPE_2D}
\title{SMED GPE 2D}
\usage{
SMED_GPE_2D(f, n0 = 10, n = 10, GP.package = "laGP",
  opt.method = "genoud", contour.fit = 0, continue.option = F,
  max.time = NULL)
}
\arguments{
\item{f}{Function}

\item{n0}{Number of initial points taken from LHS}

\item{n}{Number of points to select}

\item{GP.package}{R package to fit GP}

\item{opt.method}{Optimization method, either 'genoud' or 'LHS'.}

\item{contour.fit}{How often contour fits should be shown}

\item{continue.option}{Whether option to continue should be given}

\item{max.time}{max.time for GenSA optimization for each point.}
}
\value{
Points selected
}
\description{
SMED GPE 2D
}
\examples{
SMED_GPE_2D(TestFunctions::banana,10,10, GP.package='laGP', max.time=.01, contour.fit = 0)
}
