% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMED_select.R
\name{SMED_select}
\alias{SMED_select}
\title{SMED_select}
\usage{
SMED_select(f, n = 1, X0 = NULL, Xopt = NULL)
}
\arguments{
\item{f}{Function}

\item{n}{Number of points to select}

\item{X0}{Design matrix of already selected points (points in rows)}

\item{Xopt}{Matrix where each row is a candidate point}
}
\value{
Vector with indices of selected points
}
\description{
SMED_select
}
\examples{

}

